re-organzing examples
